Create a simple rails app
This is not a real app. The little bits of real code here are intended
to match the examples from the Scary Large Code Surgery book. This code
is intentially bad, brittle, and possibly broken.

I have included the config files and the sqlite databases so that you
can clone this repository, run `bundle`, and be ready to play with the
code. Aside from obvious development dependencies like Ruby, `sqlite3`, and
`bundler`, you shouldn't need to configure or setup anything.